Style Header, top, and middle sections
Added an approximate styling to the header, top, and middle, sections,
so they resemble the img 01 as closely as possible, without wasting
too much time around the right sizes, margins, and colors.

Rotated the original pizza image so it looks more aligned to the top
section.

The middle section might need some rework, but at the moment it
looks quite similar.